requirements.txt
aws-cdk.core
aws-cdk.aws-iam
aws-cdk.aws-ec2
aws-cdk.aws_s3
aws-cdk.aws_rds
aws-cdk.aws_ecs
aws-cdk.aws_efs
aws-cdk.aws_ses